requirements.txt
pyyaml
more-itertools
starlette==0.26.1
fastapi==0.95.1
uvicorn==0.21.1
pydantic==1.9.0
# pytube
git+https://github.com/oncename/pytube
youtube-search-python
streamlit
streamlit-keyup
wget
streamlit-authenticator==0.2.2
streamlit-extras